Refactor base package
