+++ src/main/java/com/github/sharifrahim/chatgptintegration/demo/config/OpenAIConfig.java
@@ -1,4 +1,4 @@
-package com.github.sharifrahim.chatgpt.chatgpt.integration.demo.config;
+package com.github.sharifrahim.chatgptintegration.demo.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
