+++ src/main/java/com/github/sharifrahim/chatgptintegration/demo/dto/TimelineItemDTO.java
@@ -1,4 +1,4 @@
-package com.github.sharifrahim.chatgpt.chatgpt.integration.demo.dto;
+package com.github.sharifrahim.chatgptintegration.demo.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
