+++ src/test/java/com/github/sharifrahim/chatgptintegration/demo/ChatgptIntegrationDemoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.github.sharifrahim.chatgpt.chatgpt.integration.demo;
+package com.github.sharifrahim.chatgptintegration.demo;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
